display movies as a RecyclerView items
