+++ app/src/main/res/layout/activity_main.xml
@@ -36,30 +36,31 @@
                     android:textSize="24sp"
                     android:textStyle="normal|bold" />
 
-                <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                    xmlns:tools="http://schemas.android.com/tools"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent">
 
-                    <android.support.v7.widget.RecyclerView
-                        android:id="@+id/my_recycler_view"
-                        android:layout_width="match_parent"
-                        android:layout_height="match_parent"/>
-                </LinearLayout>
-
-                <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content">
-
-                    <TableLayout
-                        android:id="@+id/main_table"
+                    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+                        xmlns:tools="http://schemas.android.com/tools"
                         android:layout_width="fill_parent"
-                        android:layout_height="fill_parent"
-                        android:layout_below="@+id/search_query"
-                        android:layout_marginTop="60dp"
-                        tools:ignore="ObsoleteLayoutParam">
-                    </TableLayout>
-                </FrameLayout>
+                        android:layout_height="wrap_content"
+                        android:layout_marginTop="60dp" >
+
+                        <android.support.v7.widget.RecyclerView
+                            android:id="@+id/my_recycler_view"
+                            android:layout_width="wrap_content"
+                            android:layout_height="wrap_content"/>
+                    </LinearLayout>
+
+                    <!--<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"-->
+                    <!--android:layout_width="wrap_content"-->
+                    <!--android:layout_height="wrap_content">-->
+
+                    <!--<TableLayout-->
+                    <!--android:id="@+id/main_table"-->
+                    <!--android:layout_width="fill_parent"-->
+                    <!--android:layout_height="fill_parent"-->
+                    <!--android:layout_below="@+id/search_query"-->
+                    <!--android:layout_marginTop="60dp"-->
+                    <!--tools:ignore="ObsoleteLayoutParam">-->
+                    <!--</TableLayout>-->
             </RelativeLayout>
         </ScrollView>
